Move BuildOperationRunner to worker shared global services
