+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/WorkerSharedGlobalScopeServices.java
@@ -48,16 +48,31 @@ import org.gradle.internal.installation.CurrentGradleInstallation;
 import org.gradle.internal.instantiation.InstantiatorFactory;
 import org.gradle.internal.logging.events.OutputEventListener;
 import org.gradle.internal.logging.progress.DefaultProgressLoggerFactory;
+import org.gradle.internal.logging.progress.ProgressLogger;
 import org.gradle.internal.logging.progress.ProgressLoggerFactory;
 import org.gradle.internal.logging.services.ProgressLoggingBridge;
 import org.gradle.internal.nativeintegration.filesystem.FileSystem;
+import org.gradle.internal.operations.BuildOperationDescriptor;
 import org.gradle.internal.operations.BuildOperationIdFactory;
+import org.gradle.internal.operations.BuildOperationListener;
+import org.gradle.internal.operations.BuildOperationListenerManager;
+import org.gradle.internal.operations.BuildOperationRunner;
+import org.gradle.internal.operations.BuildOperationState;
+import org.gradle.internal.operations.CurrentBuildOperationRef;
+import org.gradle.internal.operations.DefaultBuildOperationExecutor;
 import org.gradle.internal.operations.DefaultBuildOperationIdFactory;
+import org.gradle.internal.operations.DefaultBuildOperationRunner;
+import org.gradle.internal.operations.OperationFinishEvent;
+import org.gradle.internal.operations.OperationProgressDetails;
+import org.gradle.internal.operations.OperationProgressEvent;
+import org.gradle.internal.operations.OperationStartEvent;
 import org.gradle.internal.os.OperatingSystem;
 import org.gradle.internal.state.DefaultManagedFactoryRegistry;
 import org.gradle.internal.state.ManagedFactoryRegistry;
 import org.gradle.internal.time.Clock;
 import org.gradle.internal.time.Time;
+
+import javax.annotation.Nullable;
 
 import static org.gradle.api.internal.file.ManagedFactories.DirectoryManagedFactory;
 import static org.gradle.api.internal.file.ManagedFactories.DirectoryPropertyManagedFactory;
@@ -153,5 +168,75 @@ public class WorkerSharedGlobalScopeServices extends BasicGlobalScopeServices {
 
     ClassLoaderFactory createClassLoaderFactory() {
         return new DefaultClassLoaderFactory();
+    }
+
+    BuildOperationRunner createBuildOperationRunner(
+        Clock clock,
+        CurrentBuildOperationRef currentBuildOperationRef,
+        ProgressLoggerFactory progressLoggerFactory,
+        BuildOperationIdFactory buildOperationIdFactory,
+        BuildOperationListenerManager buildOperationListenerManager
+    ) {
+        BuildOperationListener listener = buildOperationListenerManager.getBroadcaster();
+        return new DefaultBuildOperationRunner(
+            currentBuildOperationRef,
+            clock::getCurrentTime,
+            buildOperationIdFactory,
+            () -> new ProgressListenerAdapter(listener, progressLoggerFactory, clock)
+        );
+    }
+
+    private static class ProgressListenerAdapter implements DefaultBuildOperationRunner.BuildOperationExecutionListener {
+        private final BuildOperationListener buildOperationListener;
+        private final ProgressLoggerFactory progressLoggerFactory;
+        private final Clock clock;
+        private ProgressLogger progressLogger;
+        private ProgressLogger statusProgressLogger;
+
+        public ProgressListenerAdapter(BuildOperationListener buildOperationListener, ProgressLoggerFactory progressLoggerFactory, Clock clock) {
+            this.buildOperationListener = buildOperationListener;
+            this.progressLoggerFactory = progressLoggerFactory;
+            this.clock = clock;
+        }
+
+        @Override
+        public void start(BuildOperationDescriptor descriptor, BuildOperationState operationState) {
+            buildOperationListener.started(descriptor, new OperationStartEvent(operationState.getStartTime()));
+            ProgressLogger progressLogger = progressLoggerFactory.newOperation(DefaultBuildOperationExecutor.class, descriptor);
+            this.progressLogger = progressLogger.start(descriptor.getDisplayName(), descriptor.getProgressDisplayName());
+        }
+
+        @Override
+        public void progress(BuildOperationDescriptor descriptor, String status) {
+            // Currently, need to start a new progress operation to hold the status, as changing the status of the progress operation replaces the
+            // progress display name on the console, whereas we want to display both the progress display name and the status
+            // This should be pushed down into the progress logger infrastructure so that an operation can have both a display name (that doesn't change) and
+            // a status (that does)
+            if (statusProgressLogger == null) {
+                statusProgressLogger = progressLoggerFactory.newOperation(DefaultBuildOperationExecutor.class, progressLogger);
+                statusProgressLogger.start(descriptor.getDisplayName(), status);
+            } else {
+                statusProgressLogger.progress(status);
+            }
+        }
+
+        @Override
+        public void progress(BuildOperationDescriptor descriptor, long progress, long total, String units, String status) {
+            progress(descriptor, status);
+            buildOperationListener.progress(descriptor.getId(), new OperationProgressEvent(clock.getCurrentTime(), new OperationProgressDetails(progress, total, units)));
+        }
+
+        @Override
+        public void stop(BuildOperationDescriptor descriptor, BuildOperationState operationState, @Nullable BuildOperationState parent, DefaultBuildOperationRunner.ReadableBuildOperationContext context) {
+            if (statusProgressLogger != null) {
+                statusProgressLogger.completed();
+            }
+            progressLogger.completed(context.getStatus(), context.getFailure() != null);
+            buildOperationListener.finished(descriptor, new OperationFinishEvent(operationState.getStartTime(), clock.getCurrentTime(), context.getFailure(), context.getResult()));
+        }
+
+        @Override
+        public void close(BuildOperationDescriptor descriptor, BuildOperationState operationState) {
+        }
     }
 }
